requirements.txt
flask
kafka-python
influxdb_client
certifi
